refactor(phase4a): logging / tracker helpers を library/logging_util.py に分離
train_util.py から init_trackers と LossRecorder を library/logging_util.py
へ切り出し、旧位置は re-export shim に置換。

呼び出し側はすべて train_util.init_trackers / train_util.LossRecorder
の属性アクセスで利用しているため shim 経由で動作し、修正不要。
